+++ src/features/layers-panel/layers-view.svelte.ts
@@ -5,6 +5,7 @@ interface LayerItem {
   title: string;
   visible: boolean;
   opacity: number;
+  supportsOpacity: boolean;
 }
 
 export class LayersView {